Added some casts from double to float.
svn path=/trunk/; revision=24877

+++ epan/dissectors/packet-rtcp.c
@@ -1727,12 +1727,12 @@ dissect_rtcp_xr(tvbuff_t *tvb, packet_info *pinfo _U_, int offset, proto_tree *t
             
             /* MOS LQ */
             proto_tree_add_float(content_tree, hf_rtcp_xr_voip_metrics_moslq, tvb, offset, 1, 
-                                 tvb_get_guint8(tvb, offset) / 10.0);
+                                 (float) (tvb_get_guint8(tvb, offset) / 10.0));
             offset++;
             
             /* MOS CQ */
             proto_tree_add_float(content_tree, hf_rtcp_xr_voip_metrics_moscq, tvb, offset, 1, 
-                                 tvb_get_guint8(tvb, offset) / 10.0);
+                                 (float) (tvb_get_guint8(tvb, offset) / 10.0));
             offset++;
             
             /* PLC, JB Adaptive, JB Rate */
